added guess handling function - rgb guesser game
with this function added now the code can handle how guessing works. for example, if the answer is correct the level gets higher and so does the score and a little message will display announcing it's a correct answer. The same goes for wrong answers, it'll take a life (heart) away and display a message saying the choice was incorrect. this will continue until all 3 lives are gone.